Moving fixtures to conftest.py
