+++ tests/test_arrow.py
@@ -18,53 +18,6 @@ from arrow import arrow
 from .utils import assert_datetime_equality
 
 
-@pytest.fixture(scope="class")
-def utcnow_fixture(request):
-    request.cls.arrow = arrow.Arrow.utcnow()
-
-
-@pytest.fixture(scope="class")
-def time2013_01_01_fixture(request):
-    request.cls.now = arrow.Arrow.utcnow()
-    request.cls.arrow = arrow.Arrow(2013, 1, 1)
-    request.cls.datetime = datetime(2013, 1, 1)
-
-
-@pytest.fixture(scope="class")
-def time2013_02_03_fixture(request):
-    request.cls.arrow = arrow.Arrow(2013, 2, 3, 12, 30, 45, 1)
-
-
-@pytest.fixture(scope="class")
-def time2013_02_15_fixture(request):
-    request.cls.datetime = datetime(2013, 2, 15, 3, 41, 22, 8923)
-    request.cls.arrow = arrow.Arrow.fromdatetime(request.cls.datetime)
-
-
-@pytest.fixture(scope="class")
-def timedst_fixture(request):
-    request.cls.before_1 = arrow.Arrow(
-        2016, 11, 6, 3, 59, tzinfo=tz.gettz("America/New_York")
-    )
-    request.cls.before_2 = arrow.Arrow(2016, 11, 6, tzinfo=tz.gettz("America/New_York"))
-    request.cls.after_1 = arrow.Arrow(
-        2016, 11, 6, 4, tzinfo=tz.gettz("America/New_York")
-    )
-    request.cls.after_2 = arrow.Arrow(
-        2016, 11, 6, 23, 59, tzinfo=tz.gettz("America/New_York")
-    )
-    request.cls.before_3 = arrow.Arrow(
-        2018, 11, 4, 3, 59, tzinfo=tz.gettz("America/New_York")
-    )
-    request.cls.before_4 = arrow.Arrow(2018, 11, 4, tzinfo=tz.gettz("America/New_York"))
-    request.cls.after_3 = arrow.Arrow(
-        2018, 11, 4, 4, tzinfo=tz.gettz("America/New_York")
-    )
-    request.cls.after_4 = arrow.Arrow(
-        2018, 11, 4, 23, 59, tzinfo=tz.gettz("America/New_York")
-    )
-
-
 class TestTestArrowInit:
     def test_init_bad_input(self):
